Seperate logic into different files
