+++ src/Data.js
@@ -1,0 +1,23 @@
+export const books = [
+  {
+    id: 1,
+    img:
+      "https://m.media-amazon.com/images/I/51nRHsdkmFL._AC_UL480_FMwebp_QL65_.jpg",
+    author: "Joyce N",
+    desc: "Best Author",
+  },
+  {
+    id: 2,
+    img:
+      "https://m.media-amazon.com/images/I/51nRHsdkmFL._AC_UL480_FMwebp_QL65_.jpg",
+    author: "Will N",
+    desc: "Best of Authors",
+  },
+  {
+    id: 3,
+    img:
+      "https://m.media-amazon.com/images/I/51nRHsdkmFL._AC_UL480_FMwebp_QL65_.jpg",
+    author: "Emma N",
+    desc: "Number one Authors",
+  },
+];
